search_agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
]
